Added index.js file in models folder
